pyproject.toml
[project]
name = "gtm-engine"
version = "0.1.0"
description = "Job GTM Intelligence Platform - Ingest and analyze 50k+ job postings"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    # Web framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    
    # Task queue
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",
    
    # Object storage
    "boto3>=1.34.0",
    
    # HTTP client
    "httpx>=0.26.0",
    "aiohttp>=3.9.0",
    
    # LLM clients
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    
    # Text processing
    "readability-lxml>=0.8.1",
    "lxml>=5.1.0",
    "beautifulsoup4>=4.12.0",
    
    # Validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "playwright>=1.41.0",
]

[build-system]
requires = ["setuptools>=69.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "services*", "workers*", "shared*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
